requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.26.0
lxml>=5.0.0
html5lib>=1.1
requests>=2.31.0
psycopg2-binary>=2.9.9
